% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewModels.R
\name{viewModels}
\alias{viewModels}
\title{View available models}
\usage{
viewModels()
}
\value{
table of the models in rhinno
}
\description{
Use this to view the available models in rhinno
}
\examples{
viewModels()
                          

}
